requirements.txt
streamlit
pandas
matplotlib
seaborn
scikit-learn
tensorflow
pillow
pyarrow